+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.4.0
+
+- `grep_rl/2` now routes its `opts` arg to `ls!/2` to support pre-filtering of files
+subjected to pattern matching.
+
 ## v0.3.0
 
 Adds functions for `head/2` and `tail/2`